basis.router: moved all tests for router in one test suite
